requirements.txt
beautifulsoup4==4.9.3
lxml==4.6.2
pandas==1.2.0
requests==2.25.1
streamlit==0.74.1
urllib3==1.26.2
tqdm~=4.55.1
selenium~=3.141.0
bs4~=0.0.1
PyAutoGUI~=0.9.52
img2pdf~=0.4.0
opencv-python~=4.5.1.48
Pillow~=8.1.0